requirements.txt
torch
diffusers
transformers
accelerate
sentencepiece
protobuf
pillow
gradio
